+++ frontend/src/components/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import '../styles/index.css'
 import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
